// src/api/sqlmodule/model/schemaChangeResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export interface SchemaChangeResponse {
  kind: string;
  entityType: string;
  /** @nullable */
  entityId?: string | null;
  /** @nullable */
  tempId?: string | null;
  path: string;
  description: string;
  /** @nullable */
  affectedRows?: number | null;
  severity: string;
  /** @nullable */
  code?: string | null;
}