// src/api/sqlmodule/model/schemaColumnDraft.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */
import type { SchemaColumnParameterDraft } from './schemaColumnParameterDraft';

export interface SchemaColumnDraft {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  tempId?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  physicalTypeId?: string | null;
  /** @nullable */
  isPrimaryKey?: boolean | null;
  /** @nullable */
  isRequired?: boolean | null;
  /** @nullable */
  sortOrder?: number | null;
  /** @nullable */
  parameters?: SchemaColumnParameterDraft[] | null;
}