// src/api/sqlmodule/model/schemaTableDraft.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */
import type { SchemaColumnDraft } from './schemaColumnDraft';

export interface SchemaTableDraft {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  tempId?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  sortOrder?: number | null;
  /** @nullable */
  columns?: SchemaColumnDraft[] | null;
}