// src/api/sqlmodule/model/schemaUpsertRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */
import type { SchemaRelationshipDraft } from './schemaRelationshipDraft';
import type { SchemaTableDraft } from './schemaTableDraft';

/**
 * Полное желаемое состояние схемы существующей учебной базы.
 */
export interface SchemaUpsertRequest {
  /** @nullable */
  version?: string | null;
  /** @nullable */
  tables?: SchemaTableDraft[] | null;
  /** @nullable */
  relationships?: SchemaRelationshipDraft[] | null;
}