// src/api/sqlmodule/model/targetDbSchemaResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */
import type { SchemaCapabilitiesResponse } from './schemaCapabilitiesResponse';
import type { SchemaRelationshipResponse } from './schemaRelationshipResponse';
import type { SchemaTableResponse } from './schemaTableResponse';
import type { TargetDbSchemaResponseState } from './targetDbSchemaResponseState';

/**
 * Согласованный снимок логической схемы учебной базы.
 */
export interface TargetDbSchemaResponse {
  targetDbId: string;
  dbmsId: string;
  dbName: string;
  version: string;
  state: TargetDbSchemaResponseState;
  capabilities: SchemaCapabilitiesResponse;
  tables: SchemaTableResponse[];
  relationships: SchemaRelationshipResponse[];
}